Create ProjectOutput to hold simulated data to reduce io
ProjectOutput holds a map of SimulationOutputs which were marked as
valid. Only the scenario file is kept in memory. The trajectory file
is only loaded once to see if it is valid but is not kept in memory.
SimulationOutputs can be marked as dirty. If so they will not be listed
and will be reloaded or removed with the next cleanup is done.