Add DebugGui for triangulation algorithms.
The DebugGui allows a step by step visualization of a mesh based
triangulation. The DebugGui creates a new canvas each time it is called
an thus allows different representations depending on the call location
within the algorithm.

If necessary state information on the location can be printed or a
tikzdrawing can be created.