add JoltTransformation class to allow postTransformation hooks.
This addition is needed because not all transformation can be done
with the Jolt Transformation. Each implementation of JoltTransformation
has a list of PostTransformHooks which will be executed in order on
the JsonNode representation _after_ the Jolt transformation is done.

This allows context specific transformation. If needed a postHook will
be added in the future.