add ProcessorFlag.java add MeasurementArea Check to ScenarioChecker
## ProcessorFlag.java

This is a simple String wrapped in class and used in conjunction with
the @DataProcessorClass annotation for DataProcessors. The new
processorFlags attribute of the @DataProcessorClass annotation
collects any number of Strings which are added as Flags to the
ProcessorFactory lookup table. This allows to filter Processors based
on these flags during runtime. This is used to mark all processors
which use a MeasurementArea. At runtime the scenario file
only allows access over the containing list containing
DataProcessor<?,?>. The Flags allow to filter these processors to a
set containing a specified flag.
(See DataProcessorsLinkedToMeasurementArea.java for usage)

## DataProcessor sanityCheck(Object data)

Because the DataProcessors using a MeasurementArea do not expose
this information at the generic level. This function allows to
delegate the sanityCheck to each implementation.

See AreaDensityVoronoiProcessor for usage as well as
    DataProcessorsLinkedToMeasurementArea.java

Signed-off-by: hm-schuhba1 <stefan.schuhbaeck@hm.edu>