Add ScenarioChecker for processors with  MeasurementAreas
The ProcessorFlag interface is a marker interface used anywehre a
specifc attribute of some DataProcessors are needed. For instnace
the UsesMeasurementArea Interface is implemented by all Processors
usinge one or more MeasurementAreas. In this example the interface
exposes an array of given MeasurementAreaIds which can be used
within the ScenarioChecker.

These interfaces are therefore aribtraty flags used to filter a
generic list of DataProcesors if needed.