In "DefaultModeAdapter", made resizing of scenario elements a two-step process: first, select element, then use resize handles at object edges.
Since commit c91dad12, a scenario element was selected when hovering over it.
This caused problems in the postvisualization. It was very tricky to track one
specific agent because when you moved the mouse cursor over the canvas different
agents were selected immediately. Therefore, I have reverted this behavior and
resizing a scenario is a two-step process now.