fix #280 nullpointer in gui while simulation setup
The problem occurred because the OnlineVisualization model is setup
during gui startup but was only valid after the first call to
OnlineVisualization#postUpdate where the DrawData object is created.
Only then was the OnlineVisualization valid and capable of answering
all callbacks.

This commit will fix the problem but does not fix the design the problem.
The OnlineVisualization is still created at gui startup but all
callbacks to the model are removed util the OnlineVisualization model is
fully initialized.