In ".gitlab-ci.yml", use new SSH host "webdev02-mwn.lrz.de" instead of old "webdev-mwn.lrz.de"
The new SSH host is necessary because the LRZ informed us that the
vadere.org webspace was moved to a new server (webdev02-mwn.lrz.de)
on Feb 20, 2020. The old server (webdev-mwn.lrz.de) will be shut down
in the future.

Also use proper "scp" syntax to refer to home directory on the remote host.
I.e., "./" instead of "~/", see [1].

- [1]: https://unix.stackexchange.com/questions/387859/can-i-use-scp-to-copy-files-to-the-home-directory-without-having-to-specify-it